--- SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzIntake/IntakeIO.java
@@ -1,13 +1,0 @@
-package frc.robot.CatzSubsystems.CatzIntake;
-
-import org.littletonrobotics.junction.AutoLog;
-
-import frc.robot.CatzAbstractions.io.GenericMotorIO;
-
-public interface IntakeIO extends GenericMotorIO<IntakeIO.IntakeIOInputs>{
-
-    @AutoLog
-    public static class IntakeIOInputs extends GenericMotorIO.MotorIOInputs{
-
-    }
-}
