--- SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzIntake/IntakeIOSim.java
@@ -1,7 +1,0 @@
-package frc.robot.CatzSubsystems.CatzIntake;
-
-import frc.robot.CatzAbstractions.io.GenericIOSim;
-
-public class IntakeIOSim extends GenericIOSim<IntakeIO.IntakeIOInputs> implements IntakeIO{
-
-}
