--- SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzIntake/IntakeIOTalonFX.java
@@ -1,10 +1,0 @@
-package frc.robot.CatzSubsystems.CatzIntake;
-
-import frc.robot.CatzAbstractions.io.GenericTalonFXIOReal;
-
-public class IntakeIOTalonFX extends GenericTalonFXIOReal<IntakeIO.IntakeIOInputs> implements IntakeIO{
-    public IntakeIOTalonFX(MotorIOTalonFXConfig config){
-        super(config);
-    }
-
-}
