+++ SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzIntakeRoller/IntakeRollerIOSim.java
@@ -1,0 +1,7 @@
+package frc.robot.CatzSubsystems.CatzIntakeRoller;
+
+import frc.robot.CatzAbstractions.io.GenericIOSim;
+
+public class IntakeRollerIOSim extends GenericIOSim<IntakeRollerIO.IntakeRollerIOInputs> implements IntakeRollerIO{
+
+}
