+++ SeasonCode2026/src/main/java/frc/robot/RobotContainer.java
@@ -11,8 +11,8 @@ import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import frc.robot.CatzSubsystems.CatzSuperstructure;
 import frc.robot.CatzSubsystems.CatzDriveAndRobotOrientation.CatzRobotTracker;
 import frc.robot.CatzSubsystems.CatzDriveAndRobotOrientation.Drivetrain.CatzDrivetrain;
-import frc.robot.CatzSubsystems.CatzIntake.CatzIntake;
-import frc.robot.CatzSubsystems.CatzIntake.IntakeConstants;
+import frc.robot.CatzSubsystems.CatzIntakeRoller.CatzIntakeRoller;
+import frc.robot.CatzSubsystems.CatzIntakeRoller.IntakeRollerConstants;
 import frc.robot.CatzSubsystems.CatzTurret.CatzTurret;
 import frc.robot.Commands.DriveAndRobotOrientationCmds.TeleopDriveCmd;
 
@@ -40,9 +40,9 @@ public class RobotContainer {
 
   public Command getAutonomousCommand() {
     return Commands.sequence(
-      CatzIntake.Instance.followSetpointCommand(()->IntakeConstants.setpoint).withTimeout(2.0),
+      CatzIntakeRoller.Instance.followSetpointCommand(()->IntakeRollerConstants.setpoint).withTimeout(2.0),
       Commands.waitSeconds(3),
-      CatzIntake.Instance.followSetpointCommand(()->IntakeConstants.SETPOINT2)
+      CatzIntakeRoller.Instance.followSetpointCommand(()->IntakeRollerConstants.SETPOINT2)
     );
   }
 }
